+++ checkpoint/orbax/checkpoint/_src/handlers/pytree_checkpoint_handler.py
@@ -38,13 +38,13 @@ from orbax.checkpoint import future
 from orbax.checkpoint import options as options_lib
 from orbax.checkpoint import serialization
 from orbax.checkpoint import transform_utils
-from orbax.checkpoint import type_handlers
 from orbax.checkpoint import utils
 from orbax.checkpoint._src import asyncio_utils
 from orbax.checkpoint._src.handlers import async_checkpoint_handler
 from orbax.checkpoint._src.handlers import base_pytree_checkpoint_handler
 from orbax.checkpoint._src.metadata import tree as tree_metadata
 from orbax.checkpoint._src.serialization import tensorstore_utils as ts_utils
+from orbax.checkpoint._src.serialization import type_handlers
 from orbax.checkpoint._src.tree import utils as tree_utils
 import tensorstore as ts