+++ checkpoint/orbax/checkpoint/_src/handlers/random_key_checkpoint_handler.py
@@ -24,13 +24,13 @@ from etils import epath
 import jax
 from orbax.checkpoint import checkpoint_args
 from orbax.checkpoint import future
-from orbax.checkpoint import type_handlers
 from orbax.checkpoint._src import asyncio_utils
 from orbax.checkpoint._src.handlers import array_checkpoint_handler
 from orbax.checkpoint._src.handlers import async_checkpoint_handler
 from orbax.checkpoint._src.handlers import composite_checkpoint_handler
 from orbax.checkpoint._src.handlers import json_checkpoint_handler
 from orbax.checkpoint._src.handlers import pytree_checkpoint_handler
+from orbax.checkpoint._src.serialization import type_handlers
 
 NumpyRandomKeyType = Union[tuple, dict]