+++ checkpoint/orbax/checkpoint/checkpoint_utils.py
@@ -22,11 +22,11 @@ from absl import logging
 from etils import epath
 import jax
 import numpy as np
-from orbax.checkpoint import type_handlers
 from orbax.checkpoint import utils
 from orbax.checkpoint._src.metadata import value as value_metadata
 from orbax.checkpoint._src.multihost import multihost
 from orbax.checkpoint._src.path.snapshot import snapshot as snapshot_lib
+from orbax.checkpoint._src.serialization import type_handlers
 from orbax.checkpoint.path import step as step_lib